requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-chroma
langchain-core
langchain-groq

sentence-transformers
transformers

torch
torchvision
torchaudio

youtube-transcript-api
pytube

python-dotenv
streamlit